Implemented old vs new LF

+++ backend/dataall/modules/dataset_sharing/services/share_processors/lakeformation_process_share.py
@@ -37,17 +37,19 @@ class ProcessLakeFormationShare(LFShareManager):
     def process_approved_shares(self) -> bool:
         """
         1) Grant ALL permissions to pivotRole for source database in source account
-        2) Get share principals (requester IAM role and QS groups) and build shared db name
-        3) Create the shared database in target account if it doesn't exist
+        2) Create the shared database in target account if it doesn't exist
+        3) Grant permissions to pivotRole and principals to "shared" database
         4) For each shared table:
-            a) update its status to SHARE_IN_PROGRESS with Action Start
-            b) check if share item exists on glue catalog raise error if not and flag share item status to failed
-            c) grant external account (target account) access to table -> create RAM invitation and revoke_iamallowedgroups_super_permission_from_table
-            d) accept pending RAM invitation
-            e) create resource link for table in target account
-            f) grant permission to table for requester team IAM role in source account
-            g) grant permission to resource link table for requester team IAM role in target account
-            h) update share item status to SHARE_SUCCESSFUL with Action Success
+            a) Update its status to SHARE_IN_PROGRESS with Action Start
+            b) Check if table exists on glue catalog raise error if not and flag share item status to failed
+            c) If it is a cross-account share:
+                c.1) Revoke iamallowedgroups permissions from table
+                c.2) Grant target account permissions to original table -> create RAM invitation
+                c.3) Accept pending RAM invitation
+            d) Create resource link for table in target account
+            e) grant permission to principals to shared-table in target account
+            f) grant permission to principals to resource link table
+            g) update share item status to SHARE_SUCCESSFUL with Action Success
 
         Returns
         -------
@@ -89,9 +91,10 @@ class ProcessLakeFormationShare(LFShareManager):
                     self.check_table_exists_in_source_database(share_item, table)
 
                     if self.cross_account:
-                        log.info('Processing cross-account permissions...')
+                        log.info(f'Processing cross-account permissions for table {table.GlueTableName}...')
                         # TODO: old shares, add if exists, use LFV3
-                        self.share_table_with_target_account(table)
+                        self.revoke_iam_allowed_principals_from_table(table)
+                        self.grant_target_account_permissions_to_source_table(table)
                         (
                             retry_share_table,
                             failed_invitations,
@@ -105,7 +108,7 @@ class ProcessLakeFormationShare(LFShareManager):
                         )
 
                         if retry_share_table:
-                            self.share_table_with_target_account(table)
+                            self.grant_target_account_permissions_to_source_table(table)
                             RamClient.accept_ram_invitation(
                                 source_account_id=self.source_environment.AwsAccountId,
                                 source_region=self.source_environment.region,
@@ -114,7 +117,7 @@ class ProcessLakeFormationShare(LFShareManager):
                                 source_database=self.dataset.GlueDatabaseName,
                                 source_table=table
                             )
-                    self.grant_principals_permissions_to_table_in_target(table) #TODO WITH LFV3 we might be able to remove this
+                    self.grant_principals_permissions_to_table_in_target(table)  # TODO WITH LFV3 we might be able to remove this
                     self.check_if_exists_and_create_resource_link_table_in_shared_database(table)
                     self.grant_principals_permissions_to_resource_link_table(table)
 
@@ -122,25 +125,28 @@ class ProcessLakeFormationShare(LFShareManager):
                     shared_item_SM.update_state_single_item(self.session, share_item, new_state)
 
                 except Exception as e:
-                    # must run first to ensure state transitions to failed
                     new_state = shared_item_SM.run_transition(ShareItemActions.Failure.value)
                     shared_item_SM.update_state_single_item(self.session, share_item, new_state)
                     success = False
 
-                    # statements which can throw exceptions but are not critical
                     self.handle_share_failure(table=table, error=e)
 
         return success
 
     def process_revoked_shares(self) -> bool:
         """
-        For each revoked request item:
-            a) update its status to REVOKE_IN_PROGRESS with Action Start
-            b) check if item exists on glue catalog raise error if not and flag item status to failed
-            c) revoke table resource link: undo grant permission to resource link table for team role in target account
-            d) revoke source table access: undo grant permission to table for team role in source account (and for QS Group if no other shares present for table)
-            e) delete resource link table
+        1) For each revoked table:
+            a) Update its status to REVOKE_IN_PROGRESS with Action Start
+            b) Check if table exists on glue catalog raise error if not and flag share item status to failed
+            c) Check if resource link table exists in target account
+            d) Check if the table is shared in other share requests to this target account
+            e) If c is True (resource link table exists), revoke permission to principals to resource link table
+            f) If c is True (resource link table exists), revoke permission to principals to table (and for QS Group if no other shares present for table)
+            g) If c is True and (old-share or (new-share and d is True, no other shares of this table)) then delete resource link table
+            g) If d is True (no other shares of this table with target), revoke permissions to target account to the original table
             h) update share item status to REVOKE_SUCCESSFUL with Action Success
+        2) Check if there are existing_shared_tables for this dataset with target environment
+        3) If no existing_shared_tables, delete shared database
 
         Returns
         -------
@@ -151,8 +157,6 @@ class ProcessLakeFormationShare(LFShareManager):
             '##### Starting Revoking tables #######'
         )
         success = True
-        shared_db_name = self.build_shared_db_name()
-        principals = self.get_share_principals()
         for table in self.revoked_tables:
             share_item = ShareObjectRepository.find_sharable_item(
                 self.session, self.share.shareUri, table.tableUri
@@ -163,15 +167,11 @@ class ProcessLakeFormationShare(LFShareManager):
             revoked_item_SM.update_state_single_item(self.session, share_item, new_state)
 
             try:
-                log.info(f'Starting revoke access for table: {table.GlueTableName} in database {shared_db_name} '
-                         f'For principals {principals}')
+                log.info(f'Revoking access to table: {table.GlueTableName} ')
 
                 self.check_table_exists_in_source_database(share_item, table)
-                existing = self.check_resource_link_table_exists_in_target_database(table)
-                if existing:
-                    self.revoke_principals_permissions_to_resource_link_table(table)
-                    self.revoke_principals_permissions_to_table_in_target(table)
 
+                resource_link_table_exists = self.check_resource_link_table_exists_in_target_database(table)
                 other_table_shares_in_env = True if ShareObjectRepository.other_approved_share_item_table_exists(
                     self.session,
                     self.target_environment.environmentUri,
@@ -179,37 +179,41 @@ class ProcessLakeFormationShare(LFShareManager):
                     share_item.shareItemUri
                 ) else False
 
-                if not other_table_shares_in_env:
-                    self.revoke_external_account_access_on_source_account(table)
+                if resource_link_table_exists:
+                    log.info(f'Revoking access to resource link table for: {table.GlueTableName} ')
+                    self.revoke_principals_permissions_to_resource_link_table(table)
+                    self.revoke_principals_permissions_to_table_in_target(table)
 
-                if (self.is_new_share and not other_table_shares_in_env) or not self.is_new_share:
-                    self.delete_resource_link_table_in_shared_database(table)
+                    if (self.is_new_share and not other_table_shares_in_env) or not self.is_new_share:
+                        log.info(f'Deleting resource link table for: {table.GlueTableName} ')
+                        self.delete_resource_link_table_in_shared_database(table)
+
+                if not other_table_shares_in_env:
+                    log.info(f'Revoking access from target account to table: {table.GlueTableName} ')
+                    self.revoke_external_account_access_on_source_account(table)
 
                 new_state = revoked_item_SM.run_transition(ShareItemActions.Success.value)
                 revoked_item_SM.update_state_single_item(self.session, share_item, new_state)
 
             except Exception as e:
-                # must run first to ensure state transitions to failed
                 new_state = revoked_item_SM.run_transition(ShareItemActions.Failure.value)
                 revoked_item_SM.update_state_single_item(self.session, share_item, new_state)
                 success = False
 
-                # statements which can throw exceptions but are not critical
                 self.handle_revoke_failure(table=table, error=e)
 
         try:
-            # TODO: logic to check all shares with table shares in environment
-            existing_shared_items = ShareObjectRepository.check_existing_shared_items_of_type(
-                self.session,
-                self.share.shareUri,
-                ShareableType.Table.value
+            existing_shared_tables_environment = ShareObjectRepository.list_dataset_shares_with_existing_shared_items(
+                session=self.session,
+                dataset_uri=self.dataset.datasetUri,
+                environment_uri=self.target_environment.environmentUri,
+                item_type=ShareableType.Table.value
             )
-            log.info(f'Still remaining LF resources shared = {existing_shared_items}')
-            if not existing_shared_items and self.revoked_tables:
-                log.info("Clean up LF remaining resources...")
-                clean_up_tables = self.delete_shared_database_in_target()
-                log.info(f"Clean up LF successful = {clean_up_tables}")
+
+            log.info(f'Still remaining tables shared from this dataset to this environment = {existing_shared_tables_environment}')
+            if not existing_shared_tables_environment and self.revoked_tables:
+                log.info("Deleting target shared database...")
+                self.delete_shared_database_in_target()
         except Exception as e:
             success = False
-
         return success
